refactor: OrderValidator에서 OrderTable 검증 시 Event사용 대신 Interface 도입하여 ordertable 패키지에서 처리
