+++ src/main/java/kitchenpos/menu/application/MenuService.java
@@ -5,6 +5,7 @@ import kitchenpos.menu.domain.MenuRepository;
 import kitchenpos.menu.dto.MenuRequest;
 import kitchenpos.menu.dto.MenuResponse;
 import kitchenpos.menu.exception.NoSuchMenuException;
+import kitchenpos.menu.validator.MenuValidator;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
