+++ src/test/java/kitchenpos/menu/validator/MenuValidatorTest.java
@@ -1,10 +1,11 @@
-package kitchenpos.menu.application;
+package kitchenpos.menu.validator;
 
 import kitchenpos.exception.IllegalPriceException;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuProduct;
 import kitchenpos.menu.dto.MenuProductRequest;
 import kitchenpos.menu.dto.MenuRequest;
+import kitchenpos.menu.validator.MenuValidator;
 import kitchenpos.menugroup.application.MenuGroupService;
 import kitchenpos.menugroup.domain.MenuGroup;
 import kitchenpos.menugroup.exception.NoSuchMenuGroupException;
