+++ src/test/java/kitchenpos/order/application/OrderServiceTest.java
@@ -3,7 +3,6 @@ package kitchenpos.order.application;
 import kitchenpos.menu.application.MenuService;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuProduct;
-import kitchenpos.menu.exception.NoSuchMenuException;
 import kitchenpos.menugroup.domain.MenuGroup;
 import kitchenpos.order.domain.*;
 import kitchenpos.order.dto.OrderLineItemRequest;
@@ -11,12 +10,10 @@ import kitchenpos.order.dto.OrderRequest;
 import kitchenpos.order.dto.OrderResponse;
 import kitchenpos.order.dto.OrderStatusRequest;
 import kitchenpos.order.exception.IllegalOrderException;
-import kitchenpos.order.exception.IllegalOrderLineItemException;
 import kitchenpos.order.exception.NoSuchOrderException;
+import kitchenpos.order.validator.OrderValidator;
 import kitchenpos.ordertable.application.TableService;
 import kitchenpos.ordertable.domain.OrderTable;
-import kitchenpos.ordertable.exception.IllegalOrderTableException;
-import kitchenpos.ordertable.exception.NoSuchOrderTableException;
 import kitchenpos.product.domain.Product;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -29,8 +26,6 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
-import java.time.LocalDateTime;
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
