+++ src/test/java/kitchenpos/ordertable/validator/OrderValidatorImplTest.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.application;
+package kitchenpos.ordertable.validator;
 
 import kitchenpos.menu.application.MenuService;
 import kitchenpos.menu.domain.Menu;
@@ -6,8 +6,8 @@ import kitchenpos.menu.domain.MenuProduct;
 import kitchenpos.menugroup.domain.MenuGroup;
 import kitchenpos.order.dto.OrderLineItemRequest;
 import kitchenpos.order.dto.OrderRequest;
-import kitchenpos.order.event.CreateOrderEvent;
 import kitchenpos.order.exception.IllegalOrderLineItemException;
+import kitchenpos.ordertable.application.TableService;
 import kitchenpos.ordertable.domain.OrderTable;
 import kitchenpos.ordertable.exception.IllegalOrderTableException;
 import kitchenpos.ordertable.exception.NoSuchOrderTableException;
@@ -19,7 +19,6 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.context.ApplicationEventPublisher;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -30,19 +29,18 @@ import static kitchenpos.utils.fixture.MenuProductFixtureFactory.createMenuProdu
 import static kitchenpos.utils.fixture.OrderTableFixtureFactory.createOrderTable;
 import static kitchenpos.utils.fixture.ProductFixtureFactory.createProduct;
 import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyList;
+import static org.mockito.ArgumentMatchers.anyLong;
 import static org.mockito.BDDMockito.given;
-import static org.mockito.BDDMockito.willThrow;
 
 @ExtendWith(MockitoExtension.class)
-public class OrderValidatorTest {
+public class OrderValidatorImplTest {
     @Mock
     private MenuService menuService;
     @Mock
-    private ApplicationEventPublisher eventPublisher;
+    private TableService tableService;
     @InjectMocks
-    private OrderValidator orderValidator;
+    private OrderValidatorImpl orderValidatorImpl;
 
     private MenuGroup 메뉴그룹_한식;
     private Product 김치찌개;
@@ -66,7 +64,7 @@ public class OrderValidatorTest {
     void 주문_주문테이블_검증(){
         //given
         given(menuService.findMenusByIdList(anyList())).willReturn(Arrays.asList(메뉴_김치찌개세트));
-        willThrow(NoSuchOrderTableException.class).given(eventPublisher).publishEvent(any(CreateOrderEvent.class));
+        given(tableService.findOrderTableById(anyLong())).willThrow(NoSuchOrderTableException.class);
 
         //when
         OrderRequest orderRequest = OrderRequest.of(테이블_1.getId(), Arrays.asList(
@@ -74,23 +72,24 @@ public class OrderValidatorTest {
         ));
 
         //then
-        assertThrows(NoSuchOrderTableException.class, () -> orderValidator.validate(orderRequest.toOrder()));
+        assertThrows(NoSuchOrderTableException.class, () -> orderValidatorImpl.validate(orderRequest.toOrder()));
     }
 
     @DisplayName("등록하려는 주문테이블은 비어있을 수 없다")
     @Test
     void 주문_주문테이블_Empty_검증(){
         //given
+        OrderTable 테이블_Empty = createOrderTable(2L, null, 0, true);
         given(menuService.findMenusByIdList(anyList())).willReturn(Arrays.asList(메뉴_김치찌개세트));
-        willThrow(IllegalOrderTableException.class).given(eventPublisher).publishEvent(any(CreateOrderEvent.class));
+        given(tableService.findOrderTableById(anyLong())).willReturn(테이블_Empty);
 
         //when
-        OrderRequest orderRequest = OrderRequest.of(테이블_1.getId(), Arrays.asList(
+        OrderRequest orderRequest = OrderRequest.of(테이블_Empty.getId(), Arrays.asList(
                 OrderLineItemRequest.of(메뉴_김치찌개세트.getId(), 1)
         ));
 
         //then
-        assertThrows(IllegalOrderTableException.class, () -> orderValidator.validate(orderRequest.toOrder()));
+        assertThrows(IllegalOrderTableException.class, () -> orderValidatorImpl.validate(orderRequest.toOrder()));
     }
 
     @DisplayName("주문항목은 비어있을 수 없다")
@@ -100,7 +99,7 @@ public class OrderValidatorTest {
         OrderRequest invalidOrderRequest = OrderRequest.of(테이블_1.getId(), new ArrayList<>());
 
         //then
-        assertThrows(IllegalOrderLineItemException.class, () -> orderValidator.validate(invalidOrderRequest.toOrder()));
+        assertThrows(IllegalOrderLineItemException.class, () -> orderValidatorImpl.validate(invalidOrderRequest.toOrder()));
     }
 
     @DisplayName("주문항목간 메뉴는 중복될 수 없다")
@@ -114,7 +113,7 @@ public class OrderValidatorTest {
         ));
 
         //then
-        assertThrows(IllegalOrderLineItemException.class, () -> orderValidator.validate(invalidOrderRequest.toOrder()));
+        assertThrows(IllegalOrderLineItemException.class, () -> orderValidatorImpl.validate(invalidOrderRequest.toOrder()));
     }
 
     @DisplayName("등록하려는 주문항목의 메뉴가 존재해야 한다")
@@ -130,6 +129,6 @@ public class OrderValidatorTest {
         ));
 
         //then
-        assertThrows(IllegalOrderLineItemException.class, () -> orderValidator.validate(invalidOrderRequest.toOrder()));
+        assertThrows(IllegalOrderLineItemException.class, () -> orderValidatorImpl.validate(invalidOrderRequest.toOrder()));
     }
 }
